hackthons/phyical-ai-book/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
cohere==5.11.0
qdrant-client==1.12.1
python-dotenv==1.0.1
PyJWT==2.9.0
requests==2.32.3
markdown==3.7
mistune==3.0.2
numpy==2.1.3
asyncio==3.4.3
httpx==0.27.2
beautifulsoup4==4.12.3
tiktoken==0.8.0